requirements.txt
bson==0.4.1
Django==1.9.1
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.3.2
mongoengine==0.9.0
pymongo==2.8.1
pytz==2015.7
requests==2.9.1
six==1.10.0
wheel==0.24.0
